requirements.txt
streamlit
torch
torchvision
opencv-python-headless==4.11.0.86
Pillow
numpy==1.26.4
mediapipe==0.10.21
streamlit-webrtc
